+++ src/index.js
@@ -3,19 +3,20 @@ import PropTypes from 'prop-types';
 import debounce from 'lodash.debounce';
 import memoize from 'lodash.memoize';
 import reduce from 'lodash.reduce';
+import startsWith from 'lodash.startswith';
 import classNames from 'classnames';
 
-import countryData from './country_data.js';
-import './styles.less';
+import CountryData from './CountryData.js';
 
-class ReactPhoneInput extends React.Component {
+class PhoneInput extends React.Component {
   static propTypes = {
-    excludeCountries: PropTypes.arrayOf(PropTypes.string),
+    country: PropTypes.string,
+    value: PropTypes.string,
+
     onlyCountries: PropTypes.arrayOf(PropTypes.string),
     preferredCountries: PropTypes.arrayOf(PropTypes.string),
-    defaultCountry: PropTypes.string,
+    excludeCountries: PropTypes.arrayOf(PropTypes.string),
 
-    value: PropTypes.string,
     placeholder: PropTypes.string,
     searchPlaceholder: PropTypes.string,
     disabled: PropTypes.bool,
@@ -33,12 +34,17 @@ class ReactPhoneInput extends React.Component {
     searchClass: PropTypes.string,
 
     autoFormat: PropTypes.bool,
-    disableAreaCodes: PropTypes.bool,
+
+    enableAreaCodes: PropTypes.oneOfType([
+      PropTypes.bool,
+      PropTypes.arrayOf(PropTypes.string)
+    ]),
+
     disableCountryCode: PropTypes.bool,
     disableDropdown: PropTypes.bool,
     enableLongNumbers: PropTypes.bool,
     countryCodeEditable: PropTypes.bool,
-    enableSearchField: PropTypes.bool,
+    enableSearch: PropTypes.bool,
     disableSearchIcon: PropTypes.bool,
 
     regions: PropTypes.oneOfType([
@@ -46,9 +52,19 @@ class ReactPhoneInput extends React.Component {
       PropTypes.arrayOf(PropTypes.string)
     ]),
 
-    inputExtraProps: PropTypes.object,
+    inputProps: PropTypes.object,
     localization: PropTypes.object,
     masks: PropTypes.object,
+    areaCodes: PropTypes.object,
+
+    preserveOrder: PropTypes.arrayOf(PropTypes.string),
+
+    defaultMask: PropTypes.string,
+    prefix: PropTypes.string,
+    copyNumbersOnly: PropTypes.bool,
+    renderStringAsFlag: PropTypes.string,
+    autocompleteSearch: PropTypes.bool,
+    jumpCursorToEnd: PropTypes.bool,
 
     onChange: PropTypes.func,
     onFocus: PropTypes.func,
@@ -59,13 +75,14 @@ class ReactPhoneInput extends React.Component {
   }
 
   static defaultProps = {
-    excludeCountries: [],
+    country: '',
+    value: '',
+
     onlyCountries: [],
     preferredCountries: [],
-    defaultCountry: '',
+    excludeCountries: [],
 
-    value: '',
-    placeholder: '+1 (702) 123-4567',
+    placeholder: '1 (702) 123-4567',
     searchPlaceholder: 'search',
     flagsImagePath: './flags.png',
     disabled: false,
@@ -83,24 +100,30 @@ class ReactPhoneInput extends React.Component {
     searchClass: '',
 
     autoFormat: true,
-    disableAreaCodes: false,
-    isValid: (inputNumber) => {
-      return countryData.allCountries.some((country) => {
-        return inputNumber.startsWith(country.dialCode) || country.dialCode.startsWith(inputNumber);
-      });
-    },
+    enableAreaCodes: false,
+    isValid: (inputNumber, onlyCountries) => true,
     disableCountryCode: false,
     disableDropdown: false,
     enableLongNumbers: false,
     countryCodeEditable: true,
-    enableSearchField: false,
+    enableSearch: false,
     disableSearchIcon: false,
 
     regions: '',
 
-    inputExtraProps: {},
+    inputProps: {},
     localization: {},
     masks: {},
+    areaCodes: {},
+
+    preserveOrder: [],
+
+    defaultMask: '... ... ... ... ..', // prefix+dialCode+' '+defaultMask
+    prefix: '+',
+    copyNumbersOnly: true,
+    renderStringAsFlag: '',
+    autocompleteSearch: false,
+    jumpCursorToEnd: true,
 
     onEnterKeyPress: () => {},
 
@@ -112,29 +135,22 @@ class ReactPhoneInput extends React.Component {
 
   constructor(props) {
     super(props);
-    let filteredCountries = countryData.allCountries;
-
-    if (props.regions) filteredCountries = this.filterRegions(props.regions, filteredCountries);
-    if (Object.keys(props.masks).length !== 0) filteredCountries = this.insertMasks(props.masks, filteredCountries);
-
-    const onlyCountries = this.excludeCountries(
-      this.getOnlyCountries(props.onlyCountries, filteredCountries), props.excludeCountries);
-
-    const preferredCountries = filteredCountries.filter((country) => {
-      return props.preferredCountries.some((preferredCountry) => {
-        return preferredCountry === country.iso2;
-      });
-    });
+    let { onlyCountries, preferredCountries } = new CountryData(
+      props.enableAreaCodes, props.regions,
+      props.onlyCountries, props.preferredCountries, props.excludeCountries, props.preserveOrder,
+      props.localization, props.masks, props.areaCodes,
+      props.prefix,
+    );
 
     const inputNumber = props.value.replace(/[^0-9\.]+/g, '') || '';
 
     let countryGuess;
     if (inputNumber.length > 1) {
       // Country detect by value field
-      countryGuess = this.guessSelectedCountry(inputNumber.substring(0, 6), onlyCountries, props.defaultCountry) || 0;
-    } else if (props.defaultCountry) {
+      countryGuess = this.guessSelectedCountry(inputNumber.substring(0, 6), onlyCountries, props.country) || 0;
+    } else if (props.country) {
       // Default country
-      countryGuess = onlyCountries.find(o => o.iso2 == props.defaultCountry) || 0;
+      countryGuess = onlyCountries.find(o => o.iso2 == props.country) || 0;
     } else {
       // Empty params
       countryGuess = 0;
@@ -143,7 +159,7 @@ class ReactPhoneInput extends React.Component {
     const dialCode = (
       inputNumber.length < 2 &&
       countryGuess &&
-      !inputNumber.replace(/\D/g, '').startsWith(countryGuess.dialCode)
+      !startsWith(inputNumber.replace(/\D/g, ''), countryGuess.dialCode)
     ) ? countryGuess.dialCode : '';
 
     let formattedNumber;
@@ -153,13 +169,13 @@ class ReactPhoneInput extends React.Component {
       countryGuess.name ? countryGuess.format : undefined
     );
 
-    const highlightCountryIndex = filteredCountries.findIndex(o => o == countryGuess);
+    const highlightCountryIndex = onlyCountries.findIndex(o => o == countryGuess);
 
     this.state = {
       formattedNumber,
       onlyCountries,
       preferredCountries,
-      defaultCountry: props.defaultCountry,
+      country: props.country,
       selectedCountry: countryGuess,
       highlightCountryIndex,
       queryString: '',
@@ -173,78 +189,21 @@ class ReactPhoneInput extends React.Component {
   componentDidMount() {
     if (document.addEventListener) {
       document.addEventListener('mousedown', this.handleClickOutside);
-      document.addEventListener('keydown', this.handleKeydown);
     }
   }
 
   componentWillUnmount() {
     if (document.removeEventListener) {
       document.removeEventListener('mousedown', this.handleClickOutside);
-      document.removeEventListener('keydown', this.handleKeydown);
     }
   }
 
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.defaultCountry && nextProps.defaultCountry !== this.state.defaultCountry) {
-      this.updateDefaultCountry(nextProps.defaultCountry);
+  UNSAFE_componentWillReceiveProps(nextProps) {
+    if (nextProps.country && nextProps.country !== this.state.country) {
+      this.updateCountry(nextProps.country);
     }
     else if (nextProps.value !== this.state.formattedNumber) {
       this.updateFormattedNumber(nextProps.value);
-    }
-  }
-
-  // Countries array methods
-  deleteAreaCodes = (filteredCountries) => {
-    return filteredCountries.filter((country) => {
-      return country.isAreaCode !== true;
-    });
-  }
-
-  filterRegions = (regions, filteredCountries) => {
-    if (typeof regions === 'string') {
-      const region = regions;
-      return filteredCountries.filter((country) => {
-        return country.regions.some((element) => {
-          return element === region;
-        });
-      });
-    }
-
-    return filteredCountries.filter((country) => {
-      const matches = regions.map((region) => {
-        return country.regions.some((element) => {
-          return element === region;
-        });
-      });
-      return matches.some(el => el);
-    });
-  }
-
-  insertMasks = (masks, filteredCountries) => {
-    for (let key in masks) {
-      const modifiedCountryIndex = filteredCountries.findIndex(o => o.iso2 == key);
-      if (modifiedCountryIndex != -1) filteredCountries[modifiedCountryIndex].format = masks[key]
-    }
-    return filteredCountries
-  }
-
-  getOnlyCountries = (onlyCountriesArray, filteredCountries) => {
-    if (onlyCountriesArray.length === 0) return filteredCountries;
-
-    return filteredCountries.filter((country) => {
-      return onlyCountriesArray.some((element) => {
-        return element === country.iso2;
-      });
-    });
-  }
-
-  excludeCountries = (selectedCountries, excludedCountries) => {
-    if (excludedCountries.length === 0) {
-      return selectedCountries;
-    } else {
-      return selectedCountries.filter((selCountry) => {
-        return !excludedCountries.includes(selCountry.iso2);
-      });
     }
   }
 
@@ -254,17 +213,17 @@ class ReactPhoneInput extends React.Component {
     }
     // don't include the preferred countries in search
     const probableCountries = this.state.onlyCountries.filter((country) => {
-      return country.name.toLowerCase().startsWith(queryString.toLowerCase());
+      return startsWith(country.name.toLowerCase(), queryString.toLowerCase());
     }, this);
     return probableCountries[0];
   });
 
-  guessSelectedCountry = memoize((inputNumber, onlyCountries, defaultCountry) => {
-    const secondBestGuess = onlyCountries.find(o => o.iso2 == defaultCountry) || {};
+  guessSelectedCountry = memoize((inputNumber, onlyCountries, country) => {
+    const secondBestGuess = onlyCountries.find(o => o.iso2 == country);
     if (inputNumber.trim() === '') return secondBestGuess;
 
     const bestGuess = onlyCountries.reduce((selectedCountry, country) => {
-      if (inputNumber.startsWith(country.dialCode)) {
+      if (startsWith(inputNumber, country.dialCode)) {
         if (country.dialCode.length > selectedCountry.dialCode.length) {
           return country;
         }
@@ -280,49 +239,55 @@ class ReactPhoneInput extends React.Component {
   });
 
   // Hooks for updated props
-  updateDefaultCountry = (country) => {
-    const newSelectedCountry = this.state.onlyCountries.find(o => o.iso2 == country);
-    this.setState({
-      defaultCountry: country,
-      selectedCountry: newSelectedCountry,
-      formattedNumber: this.props.disableCountryCode ? '' : '+' + newSelectedCountry.dialCode
-    });
+  updateCountry = (country) => {
+    let newSelectedCountry;
+    if (country.indexOf(0) >= '0' && country.indexOf(0) <= '9') {
+      newSelectedCountry = this.state.onlyCountries.find(o => o.dialCode == +country);
+    } else {
+      newSelectedCountry = this.state.onlyCountries.find(o => o.iso2 == country);
+    }
+    if (newSelectedCountry && newSelectedCountry.dialCode) {
+      this.setState({
+        country: country,
+        selectedCountry: newSelectedCountry,
+        formattedNumber: this.props.disableCountryCode ? '' : this.props.prefix+newSelectedCountry.dialCode
+      });
+    }
   }
 
   updateFormattedNumber(value) {
-    const { onlyCountries, defaultCountry } = this.state;
+    const { onlyCountries, country } = this.state;
     let newSelectedCountry;
     let inputNumber = value;
     let formattedNumber = value;
 
-    // if inputNumber does not start with '+', then use default country's dialing prefix,
+    // if inputNumber does not start with this.props.prefix, then use default country's dialing prefix,
     // otherwise use logic for finding country based on country prefix.
-    if (!inputNumber.startsWith('+')) {
-      newSelectedCountry = this.state.selectedCountry || onlyCountries.find(o => o.iso2 == defaultCountry);
-      const dialCode = newSelectedCountry && !inputNumber.replace(/\D/g, '').startsWith(newSelectedCountry.dialCode) ? newSelectedCountry.dialCode : '';
+    if (!startsWith(inputNumber, this.props.prefix)) {
+      newSelectedCountry = this.state.selectedCountry || onlyCountries.find(o => o.iso2 == country);
+      const dialCode = newSelectedCountry && !startsWith(inputNumber.replace(/\D/g, ''), newSelectedCountry.dialCode) ? newSelectedCountry.dialCode : '';
       formattedNumber = this.formatNumber(
         (this.props.disableCountryCode ? '' : dialCode) + inputNumber.replace(/\D/g, ''),
-        newSelectedCountry ? newSelectedCountry.format : undefined
+        newSelectedCountry ? (newSelectedCountry.format || this.getDefaultMask(newSelectedCountry)) : undefined
       );
     }
     else {
       inputNumber = inputNumber.replace(/\D/g, '');
-      newSelectedCountry = this.guessSelectedCountry(inputNumber.substring(0, 6), onlyCountries, defaultCountry);
-      formattedNumber = this.formatNumber(inputNumber, newSelectedCountry.format);
+      newSelectedCountry = this.guessSelectedCountry(inputNumber.substring(0, 6), onlyCountries, country) || this.state.selectedCountry;
+      formattedNumber = newSelectedCountry ?
+        this.formatNumber(inputNumber, newSelectedCountry.format || this.getDefaultMask(newSelectedCountry)) : inputNumber;
     }
 
     this.setState({ selectedCountry: newSelectedCountry, formattedNumber });
   }
 
+  getDefaultMask = country => this.props.prefix+''.padEnd(country.dialCode.length,'.')+' '+this.props.defaultMask;
+
   // View methods
   scrollTo = (country, middle) => {
-    if (!country)
-      return;
-
+    if (!country) return;
     const container = this.dropdownRef;
-
-    if (!container || !document.body)
-      return;
+    if (!container || !document.body) return;
 
     const containerHeight = container.offsetHeight;
     const containerOffset = container.getBoundingClientRect();
@@ -339,7 +304,7 @@ class ReactPhoneInput extends React.Component {
     let newScrollTop = elementTop - containerTop + container.scrollTop;
     const middleOffset = (containerHeight / 2) - (elementHeight / 2);
 
-    if (elementTop < containerTop) {
+    if (this.props.enableSearch ? elementTop < containerTop + 32 : elementTop < containerTop) {
       // scroll up
       if (middle) {
         newScrollTop -= middleOffset;
@@ -356,6 +321,12 @@ class ReactPhoneInput extends React.Component {
     }
   }
 
+  scrollToTop = () => {
+    const container = this.dropdownRef;
+    if (!container || !document.body) return;
+    container.scrollTop = 0;
+  }
+
   formatNumber = (text, patternArg) => {
     const { disableCountryCode, enableLongNumbers, autoFormat } = this.props;
 
@@ -369,13 +340,13 @@ class ReactPhoneInput extends React.Component {
     }
 
     if (!text || text.length === 0) {
-      return disableCountryCode ? '' : '+';
+      return disableCountryCode ? '' : this.props.prefix;
     }
 
     // for all strings with length less than 3, just return it (1, 2 etc.)
     // also return the same text if the selected country has no fixed format
     if ((text && text.length < 2) || !pattern || !autoFormat) {
-      return disableCountryCode ? text : `+${text}`;
+      return disableCountryCode ? text : this.props.prefix+text;
     }
 
     const formattedObject = reduce(pattern, (acc, character) => {
@@ -431,48 +402,46 @@ class ReactPhoneInput extends React.Component {
     return {
       name: this.state.selectedCountry.name || '',
       dialCode: this.state.selectedCountry.dialCode || '',
-      countryCode: this.state.selectedCountry.iso2 || ''
+      countryCode: this.state.selectedCountry.iso2 || '',
+      format: this.state.selectedCountry.format || ''
     }
   }
 
   handleFlagDropdownClick = () => {
     if (!this.state.showDropdown && this.props.disabled) return;
+    const { preferredCountries, selectedCountry } = this.state
+    const allCountries = preferredCountries.concat(this.state.onlyCountries)
 
-    if (this.state.preferredCountries.includes(this.state.selectedCountry)) {
-      this.setState({
-        showDropdown: !this.state.showDropdown,
-        highlightCountryIndex: this.state.preferredCountries.findIndex(o => o == this.state.selectedCountry)
-      }, () => {
-        if (this.state.showDropdown) {
-          this.scrollTo(this.getElement(this.state.highlightCountryIndex));
-        }
-      });
+    let highlightCountryIndex;
+    if (preferredCountries.includes(selectedCountry)) {
+      highlightCountryIndex = preferredCountries.findIndex(o => o == selectedCountry)
+    } else {
+      highlightCountryIndex = this.props.enableAreaCodes ? allCountries.findIndex(o => o == selectedCountry)
+        : allCountries.findIndex(o => o.iso2 == selectedCountry.iso2)
     }
-    else {
-      const onlyCountries = this.props.disableAreaCodes ? this.deleteAreaCodes(this.state.onlyCountries) : this.state.onlyCountries;
 
-      this.setState({
-        showDropdown: !this.state.showDropdown,
-        highlightCountryIndex: this.props.disableAreaCodes ? onlyCountries.findIndex(o => o.iso2 == this.state.selectedCountry.iso2) :
-          onlyCountries.findIndex(o => o == this.state.selectedCountry)
-      }, () => {
-        if (this.state.showDropdown) {
-          this.scrollTo(this.getElement(this.state.highlightCountryIndex + this.state.preferredCountries.length));
-        }
-      });
-    }
+    this.setState({
+      showDropdown: !this.state.showDropdown,
+      highlightCountryIndex,
+    }, () => {
+      if (this.state.showDropdown) {
+        this.scrollTo(this.getElement(this.state.highlightCountryIndex));
+      }
+    });
   }
 
   handleInput = (e) => {
-    let formattedNumber = this.props.disableCountryCode ? '' : '+';
+    let formattedNumber = this.props.disableCountryCode ? '' : this.props.prefix;
     let newSelectedCountry = this.state.selectedCountry;
     let freezeSelection = this.state.freezeSelection;
 
-    if(!this.props.countryCodeEditable) {
-        const updatedInput = '+' + newSelectedCountry.dialCode;
-        if (e.target.value.length < updatedInput.length) {
-            return;
-        }
+    if (!this.props.countryCodeEditable) {
+      const mainCode = newSelectedCountry.hasAreaCodes ?
+        this.state.onlyCountries.find(o => o.iso2 === newSelectedCountry.iso2 && o.mainCode).dialCode :
+        newSelectedCountry.dialCode;
+
+      const updatedInput = this.props.prefix+mainCode;
+      if (e.target.value.slice(0, updatedInput.length) !== updatedInput) return;
     }
 
     // Does not exceed 15 digit phone number limit
@@ -496,10 +465,11 @@ class ReactPhoneInput extends React.Component {
       // the guess country function can then use memoization much more effectively since the set of input it
       // gets has drastically reduced
       if (!this.state.freezeSelection || this.state.selectedCountry.dialCode.length > inputNumber.length) {
-        newSelectedCountry = this.guessSelectedCountry(inputNumber.substring(0, 6), this.state.onlyCountries, this.state.defaultCountry);
+        newSelectedCountry = this.guessSelectedCountry(inputNumber.substring(0, 6), this.state.onlyCountries, this.state.country) || this.state.selectedCountry;
         freezeSelection = false;
       }
-      formattedNumber = this.formatNumber(inputNumber, newSelectedCountry.format); // remove all non numerals from the input
+      formattedNumber = newSelectedCountry ?
+        this.formatNumber(inputNumber, newSelectedCountry.format || this.getDefaultMask(newSelectedCountry)) : inputNumber;
       newSelectedCountry = newSelectedCountry.dialCode ? newSelectedCountry : this.state.selectedCountry;
     }
 
@@ -525,7 +495,7 @@ class ReactPhoneInput extends React.Component {
         this.numberInputRef.setSelectionRange(caretPosition, caretPosition);
       }
 
-      if (this.props.onChange) this.props.onChange(this.state.formattedNumber, this.getCountryData());
+      if (this.props.onChange) this.props.onChange(this.state.formattedNumber, this.getCountryData(), e);
     });
   }
 
@@ -534,17 +504,23 @@ class ReactPhoneInput extends React.Component {
     if (this.props.onClick) this.props.onClick(e, this.getCountryData());
   }
 
+  handleDoubleClick = (e) => {
+    const len = e.target.value.length;
+    e.target.setSelectionRange(0, len);
+  }
+
   handleFlagItemClick = (country) => {
     const currentSelectedCountry = this.state.selectedCountry;
-    const nextSelectedCountry = this.state.onlyCountries.find(o => o == country);
+    const newSelectedCountry = this.state.onlyCountries.find(o => o == country);
+    if (!newSelectedCountry) return;
 
     const unformattedNumber = this.state.formattedNumber.replace(' ', '').replace('(', '').replace(')', '').replace('-', '');
-    const newNumber = unformattedNumber.length > 1 ? unformattedNumber.replace(currentSelectedCountry.dialCode, nextSelectedCountry.dialCode) : nextSelectedCountry.dialCode;
-    const formattedNumber = this.formatNumber(newNumber.replace(/\D/g, ''), nextSelectedCountry.format);
+    const newNumber = unformattedNumber.length > 1 ? unformattedNumber.replace(currentSelectedCountry.dialCode, newSelectedCountry.dialCode) : newSelectedCountry.dialCode;
+    const formattedNumber = this.formatNumber(newNumber.replace(/\D/g, ''), newSelectedCountry.format || this.getDefaultMask(newSelectedCountry));
 
     this.setState({
       showDropdown: false,
-      selectedCountry: nextSelectedCountry,
+      selectedCountry: newSelectedCountry,
       freezeSelection: true,
       formattedNumber
     }, () => {
@@ -556,22 +532,29 @@ class ReactPhoneInput extends React.Component {
   handleInputFocus = (e) => {
     // if the input is blank, insert dial code of the selected country
     if (this.numberInputRef) {
-      if (this.numberInputRef.value === '+' && this.state.selectedCountry && !this.props.disableCountryCode) {
+      if (this.numberInputRef.value === this.props.prefix && this.state.selectedCountry && !this.props.disableCountryCode) {
         this.setState({
-          formattedNumber: '+' + this.state.selectedCountry.dialCode
-        }, () => setTimeout(this.cursorToEnd, 10));
+          formattedNumber: this.props.prefix + this.state.selectedCountry.dialCode
+        }, () => {this.props.jumpCursorToEnd && setTimeout(this.cursorToEnd, 0)});
       }
     }
 
     this.setState({ placeholder: '' });
 
     this.props.onFocus && this.props.onFocus(e, this.getCountryData());
-    setTimeout(this.cursorToEnd, 10);
+    this.props.jumpCursorToEnd && setTimeout(this.cursorToEnd, 0);
   }
 
   handleInputBlur = (e) => {
     if (!e.target.value) this.setState({ placeholder: this.props.placeholder });
     this.props.onBlur && this.props.onBlur(e, this.getCountryData());
+  }
+
+  handleInputCopy = (e) => {
+    if (!this.props.copyNumbersOnly) return;
+    const text = window.getSelection().toString().replace(/[^0-9]+/g,'');
+    e.clipboardData.setData('text/plain', text);
+    e.preventDefault();
   }
 
   getHighlightCountryIndex = (direction) => {
@@ -582,6 +565,7 @@ class ReactPhoneInput extends React.Component {
       return highlightCountryIndex - direction;
     }
 
+    if (this.props.enableSearch && highlightCountryIndex > this.getSearchFilteredCountries().length) return 0; // select first country
     return highlightCountryIndex;
   }
 
@@ -596,24 +580,31 @@ class ReactPhoneInput extends React.Component {
 
   handleKeydown = (e) => {
     const { keys } = this.props;
-    if (!this.state.showDropdown || this.props.disabled) return;
     const { target: { id } } = e;
-    if (id === 'search-box') return; // don't process events coming from the search field
+
+    if (id === 'flag-dropdown' && e.which === keys.ENTER && !this.state.showDropdown) return this.handleFlagDropdownClick();
+    if (id === 'phone-form-control' && (e.which === keys.ENTER || e.which === keys.ESC)) return e.target.blur();
+
+    if (!this.state.showDropdown || this.props.disabled) return;
+    if (id === 'search-box') {
+      if (e.which !== keys.UP && e.which !== keys.DOWN && e.which !== keys.ENTER) {
+        if (e.which === keys.ESC && e.target.value === '') {
+         // do nothing // if search field is empty, pass event (close dropdown)
+       } else {
+         return; // don't process other events coming from the search field
+       }
+      }
+    }
 
     // ie hack
-    if (e.preventDefault) {
-      e.preventDefault();
-    } else {
-      e.returnValue = false;
-    }
+    if (e.preventDefault) { e.preventDefault(); }
+    else { e.returnValue = false; }
 
     const moveHighlight = (direction) => {
       this.setState({
         highlightCountryIndex: this.getHighlightCountryIndex(direction)
       }, () => {
-        this.scrollTo(this.getElement(
-          this.state.highlightCountryIndex + this.state.preferredCountries.length
-        ), true);
+        this.scrollTo(this.getElement(this.state.highlightCountryIndex), true);
       });
     }
 
@@ -625,7 +616,11 @@ class ReactPhoneInput extends React.Component {
         moveHighlight(-1);
         break;
       case keys.ENTER:
-        this.handleFlagItemClick(this.state.onlyCountries[this.state.highlightCountryIndex], e);
+        if (this.props.enableSearch) {
+          this.handleFlagItemClick(this.getSearchFilteredCountries()[this.state.highlightCountryIndex] || this.getSearchFilteredCountries()[0], e);
+        } else {
+          this.handleFlagItemClick([...this.state.preferredCountries, ...this.state.onlyCountries][this.state.highlightCountryIndex], e);
+        }
         break;
       case keys.ESC:
         this.setState({
@@ -658,41 +653,53 @@ class ReactPhoneInput extends React.Component {
 
   handleSearchChange = (e) => {
     const { currentTarget: { value: searchValue } } = e;
-    this.setState({ searchValue });
+    const { preferredCountries, selectedCountry } = this.state
+    let highlightCountryIndex = 0;
+
+    if (searchValue === '' && selectedCountry) {
+      const { onlyCountries } = this.state
+      highlightCountryIndex = preferredCountries.concat(onlyCountries).findIndex(o => o == selectedCountry);
+      // wait asynchronous search results re-render, then scroll
+      setTimeout(() => this.scrollTo(this.getElement(highlightCountryIndex)), 100)
+    }
+    this.setState({ searchValue, highlightCountryIndex });
   }
 
   getDropdownCountryName = (country) => {
-    if (this.props.localization[country.name] != undefined) {
-      return this.props.localization[country.name];
+    return country.localName || country.name;
+  }
+
+  getSearchFilteredCountries = () => {
+    const { preferredCountries, onlyCountries, searchValue } = this.state
+    const { enableSearch } = this.props
+    const allCountries = preferredCountries.concat(onlyCountries);
+    const sanitizedSearchValue = searchValue.trim().toLowerCase();
+    if (enableSearch && sanitizedSearchValue) {
+      // [...new Set()] to get rid of duplicates
+      // firstly search by iso2 code
+      const iso2countries = allCountries.filter(({ name, localName, iso2, dialCode }) =>
+        [`${iso2}`].some(field => field.toLowerCase().includes(sanitizedSearchValue)))
+      const searchedCountries = allCountries.filter(({ name, localName, iso2, dialCode }) =>
+        [`${name}`, `${localName}`, this.props.prefix+dialCode].some(field => field.toLowerCase().includes(sanitizedSearchValue)))
+      this.scrollToTop()
+      return [...new Set([].concat(iso2countries, searchedCountries))]
+    } else {
+      return allCountries
     }
-    else if (this.props.localization[country.iso2] != undefined) {
-      return this.props.localization[country.iso2];
-    }
-    return country.name;
   }
 
   getCountryDropdownList = () => {
-    const { preferredCountries, onlyCountries, highlightCountryIndex, showDropdown, searchValue } = this.state;
-    const { enableSearchField, disableSearchIcon, searchClass, searchStyle, searchPlaceholder } = this.props;
+    const { preferredCountries, highlightCountryIndex, showDropdown, searchValue } = this.state;
+    const { enableSearch, disableSearchIcon, searchClass, searchStyle, searchPlaceholder, autocompleteSearch } = this.props;
 
-    const countryIsPreferred = this.state.preferredCountries.includes(this.state.selectedCountry);
-    const allCountries = preferredCountries.concat(onlyCountries);
+    const searchedCountries = this.getSearchFilteredCountries()
 
-    const sanitizedSearchValue = searchValue.trim().toLowerCase();
-    let filteredCountries = (enableSearchField && sanitizedSearchValue)
-      // using [...new Set()] here to get rid of duplicates
-      ? [...new Set(allCountries.filter(({ name, iso2, dialCode }) =>
-        [`${name}`, `${iso2}`, `+${dialCode}`].some(field => field.toLowerCase().includes(sanitizedSearchValue))))]
-      : allCountries;
-
-    if (this.props.disableAreaCodes) filteredCountries = this.deleteAreaCodes(filteredCountries);
-
-    let countryDropdownList = filteredCountries.map((country, index) => {
+    let countryDropdownList = searchedCountries.map((country, index) => {
       const itemClasses = classNames({
         country: true,
         preferred: country.iso2 === 'us' || country.iso2 === 'gb',
         active: country.iso2 === 'us',
-        highlight: countryIsPreferred ? highlightCountryIndex === index : highlightCountryIndex === index - preferredCountries.length
+        highlight: highlightCountryIndex === index
       });
 
       const inputFlagClasses = `flag ${country.iso2}`;
@@ -704,19 +711,20 @@ class ReactPhoneInput extends React.Component {
           data-flag-key={`flag_no_${index}`}
           className={itemClasses}
           data-dial-code='1'
+          tabIndex='0'
           data-country-code={country.iso2}
           onClick={() => this.handleFlagItemClick(country)}
         >
           <div className={inputFlagClasses}/>
           <span className='country-name'>{this.getDropdownCountryName(country)}</span>
-          <span className='dial-code'>{'+' + country.dialCode}</span>
+          <span className='dial-code'>{country.format ? this.formatNumber(country.dialCode, country.format) : (this.props.prefix+country.dialCode)}</span>
         </li>
       );
     });
 
     const dashedLi = (<li key={'dashes'} className='divider'/>);
     // let's insert a dashed line in between preffered countries and the rest
-    (preferredCountries.length > 0) &&
+    (preferredCountries.length > 0) && (!enableSearch || enableSearch && !searchValue.trim()) &&
     countryDropdownList.splice(preferredCountries.length, 0, dashedLi);
 
     const dropDownClasses = classNames({
@@ -731,7 +739,7 @@ class ReactPhoneInput extends React.Component {
         className={dropDownClasses}
         style={this.props.dropdownStyle}
       >
-        {enableSearchField && (
+        {enableSearch && (
           <li
             className={classNames({
               search: true,
@@ -759,6 +767,7 @@ class ReactPhoneInput extends React.Component {
               type='search'
               placeholder={searchPlaceholder}
               autoFocus={true}
+              autoComplete={autocompleteSearch ? 'on' : 'off'}
               value={searchValue}
               onChange={this.handleSearchChange}
             />
@@ -776,58 +785,71 @@ class ReactPhoneInput extends React.Component {
   }
 
   render() {
-    const { selectedCountry, showDropdown, formattedNumber } = this.state;
-    const disableDropdown = this.props.disableDropdown;
+    const { onlyCountries, selectedCountry, showDropdown, formattedNumber } = this.state;
+    const { disableDropdown, renderStringAsFlag } = this.props;
 
     const arrowClasses = classNames({'arrow': true, 'up': showDropdown});
     const inputClasses = classNames({
       [this.props.inputClass]: true,
       'form-control': true,
-      'invalid-number': !this.props.isValid(formattedNumber.replace(/\D/g, ''))
+      'invalid-number': !this.props.isValid(formattedNumber.replace(/\D/g, ''), onlyCountries),
+      'open': showDropdown,
     });
-
+    const selectedFlagClasses = classNames({
+      'selected-flag': true,
+      'open': showDropdown,
+    });
     const flagViewClasses = classNames({
       [this.props.buttonClass]: true,
       'flag-dropdown': true,
-      'open-dropdown': showDropdown
+      'open': showDropdown,
     });
     const inputFlagClasses = `flag ${selectedCountry && selectedCountry.iso2}`;
 
     return (
       <div
         className={this.props.containerClass}
-        style={this.props.containerStyle}>
+        style={this.props.style || this.props.containerStyle}
+        onKeyDown={this.handleKeydown}>
         <input
           className={inputClasses}
+          id='phone-form-control'
           style={this.props.inputStyle}
           onChange={this.handleInput}
           onClick={this.handleInputClick}
+          onDoubleClick={this.handleDoubleClick}
           onFocus={this.handleInputFocus}
           onBlur={this.handleInputBlur}
+          onCopy={this.handleInputCopy}
           value={formattedNumber}
           ref={el => this.numberInputRef = el}
           onKeyDown={this.handleInputKeyDown}
           placeholder={this.props.placeholder}
           disabled={this.props.disabled}
           type='tel'
-          {...this.props.inputExtraProps}
+          {...this.props.inputProps}
         />
 
         <div
           className={flagViewClasses}
+          id='flag-dropdown'
           style={this.props.buttonStyle}
-          onKeyDown={this.handleKeydown}
           ref={el => this.dropdownContainerRef = el}
+          tabIndex='0'
+          role='button'
         >
+          {renderStringAsFlag ?
+          <div className={selectedFlagClasses}>{renderStringAsFlag}</div>
+          :
           <div
             onClick={disableDropdown ? undefined : this.handleFlagDropdownClick}
-            className='selected-flag'
+            className={selectedFlagClasses}
             title={selectedCountry ? `${selectedCountry.name}: + ${selectedCountry.dialCode}` : ''}
           >
             <div className={inputFlagClasses}>
               {!disableDropdown && <div className={arrowClasses}></div>}
             </div>
-          </div>
+          </div>}
 
           {showDropdown && this.getCountryDropdownList()}
         </div>
@@ -836,4 +858,4 @@ class ReactPhoneInput extends React.Component {
   }
 }
 
-export default ReactPhoneInput;
+export default PhoneInput;